+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - [8713](https://github.com/vegaprotocol/vega/issues/8713) - Wallet name with upper-case letter is allowed, again
 - [8698](https://github.com/vegaprotocol/vega/issues/8698) - Always set the `ToAccount` field when clearing fees.
 - [8711](https://github.com/vegaprotocol/vega/issues/8711) - Fix market lineage trigger.
+- [8737](https://github.com/vegaprotocol/vega/issues/8737) - Fix `GetStopOrder` errors when order has a state change.
 
 
 ## 0.72.0